+++ react_on_rails_pro/packages/node-renderer/src/worker/vm.ts
@@ -57,6 +57,13 @@ declare global {
   var ReactOnRails: ROR | undefined;
 }
 
+const extendContext = (contextObject: vm.Context, additionalContext: Record<string, unknown>) => {
+  if (log.level === 'debug') {
+    log.debug(`Adding ${Object.keys(additionalContext).join(', ')} to context object.`);
+  }
+  Object.assign(contextObject, additionalContext);
+};
+
 export async function buildVM(filePath: string) {
   if (filePath === vmBundleFilePath && context) {
     return Promise.resolve(true);
@@ -69,10 +76,7 @@ export async function buildVM(filePath: string) {
     sharedConsoleHistory = new SharedConsoleHistory();
     const contextObject = { sharedConsoleHistory };
     if (supportModules) {
-      log.debug(
-        'Adding Buffer, process, setTimeout, setInterval, setImmediate, clearTimeout, clearInterval, clearImmediate to context object.',
-      );
-      Object.assign(contextObject, {
+      extendContext(contextObject, {
         Buffer,
         process,
         setTimeout,
@@ -81,13 +85,12 @@ export async function buildVM(filePath: string) {
         clearTimeout,
         clearInterval,
         clearImmediate,
+        queueMicrotask,
       });
     }
 
     if (additionalContextIsObject) {
-      const keysString = Object.keys(additionalContext).join(', ');
-      log.debug(`Adding ${keysString} to context object.`);
-      Object.assign(contextObject, additionalContext);
+      extendContext(contextObject, additionalContext);
     }
     context = vm.createContext(contextObject);
 
@@ -142,6 +145,7 @@ export async function buildVM(filePath: string) {
       vm.runInContext(`function clearTimeout() {}`, context);
       vm.runInContext(`function clearInterval() {}`, context);
       vm.runInContext(`function clearImmediate() {}`, context);
+      vm.runInContext(`function queueMicrotask() {}`, context);
     }
 
     // Run bundle code in created context: